+++ rotkehlchen/tests/unit/test_tokens.py
@@ -1,23 +1,27 @@
 import datetime
+from contextlib import suppress
 from typing import TYPE_CHECKING, Any
 from unittest.mock import MagicMock, patch
 
 import gevent
 import pytest
 
-from rotkehlchen.assets.utils import _query_or_get_given_token_info
+from rotkehlchen.assets.utils import _query_or_get_given_token_info, get_or_create_evm_token
 from rotkehlchen.chain.ethereum.tokens import EthereumTokens
 from rotkehlchen.chain.evm.tokens import generate_multicall_chunks
 from rotkehlchen.chain.evm.types import string_to_evm_address
 from rotkehlchen.constants import ONE
 from rotkehlchen.constants.assets import A_OMG, A_WETH
+from rotkehlchen.errors.misc import InputError
 from rotkehlchen.fval import FVal
+from rotkehlchen.globaldb.handler import GlobalDBHandler
 from rotkehlchen.tests.utils.constants import A_LPT
 from rotkehlchen.tests.utils.factories import make_evm_address
 from rotkehlchen.types import ChainID, ChecksumEvmAddress, EvmTokenKind, SupportedBlockchain
 from rotkehlchen.utils.misc import ts_now
 
 if TYPE_CHECKING:
+    from rotkehlchen.chain.ethereum.node_inquirer import EthereumInquirer
     from rotkehlchen.db.dbhandler import DBHandler
 
 
@@ -261,3 +265,46 @@ def test_flaky_binding_parameter_zero(
     with database.conn.read_ctx() as cursor:
         for address in ethereum_accounts:
             database.get_tokens_for_address(cursor, address, SupportedBlockchain.ETHEREUM)
+
+
+@pytest.mark.parametrize('number_of_eth_accounts', [1])
+def test_old_curve_gauge(ethereum_inquirer: 'EthereumInquirer'):
+    """Test that querying new and old gauges get the data correctly
+    Old one should peak the default values provided and the new one should
+    get values on chain
+    """
+    # old gauge
+    gauge_address = string_to_evm_address('0xC2b1DF84112619D190193E48148000e3990Bf627')
+    with suppress(InputError):
+        GlobalDBHandler.delete_evm_token(gauge_address, chain_id=ChainID.ETHEREUM)
+
+    gauge_token = get_or_create_evm_token(
+        userdb=ethereum_inquirer.database,
+        evm_address=gauge_address,
+        chain_id=ethereum_inquirer.chain_id,
+        evm_inquirer=ethereum_inquirer,
+        decimals=18,
+        name='USDK Gauge Deposit',
+        symbol='USDK curve-gauge',
+    )
+    assert gauge_token.name == 'USDK Gauge Deposit'
+    assert gauge_token.symbol == 'USDK curve-gauge'
+    assert gauge_token.decimals == 18
+
+    # new gauge
+    gauge_address = string_to_evm_address('0x182B723a58739a9c974cFDB385ceaDb237453c28')
+    with suppress(InputError):
+        GlobalDBHandler.delete_evm_token(gauge_address, chain_id=ChainID.ETHEREUM)
+
+    gauge_token = get_or_create_evm_token(
+        userdb=ethereum_inquirer.database,
+        evm_address=gauge_address,
+        chain_id=ethereum_inquirer.chain_id,
+        evm_inquirer=ethereum_inquirer,
+        fallback_decimals=18,
+        fallback_name='stETH Gauge Deposit',
+        fallback_symbol='stETH curve-gauge',
+    )
+    assert gauge_token.name == 'Curve.fi steCRV Gauge Deposit'
+    assert gauge_token.symbol == 'steCRV-gauge'
+    assert gauge_token.decimals == 18